consolidate site analytics client
